+++ src/tests/conduit/fortran/t_f_conduit_node_obj.f90
@@ -893,6 +893,34 @@ contains
         call conduit_node_obj_destroy(n)
 
     end subroutine t_node_obj_names_embedded_slashes
+    
+    !--------------------------------------------------------------------------
+     subroutine t_node_obj_fetch_existing
+         type(node) n
+         type(node) n_1
+         type(node) n_1_test
+         real(kind=8) val
+
+
+         !----------------------------------------------------------------------
+         call set_case_name("t_node_obj_fetch_existing")
+         !----------------------------------------------------------------------
+
+         n = conduit_node_obj_create()
+
+         n_1 = n%fetch("normal/path")
+         call n_1%set(10.0d+0)
+         val = n_1%as_float64();
+         call assert_equals(10.0d+0, val)
+
+         n_1_test = n%fetch_existing("normal/path")
+
+         val = n_1_test%as_float64();
+         call assert_equals(10.0d+0, val)
+    
+         call conduit_node_obj_destroy(n)
+
+     end subroutine t_node_obj_fetch_existing
 
 
 !------------------------------------------------------------------------------
@@ -933,6 +961,7 @@ program fortran_test
   call t_node_obj_parse
   call t_node_obj_save_load
   call t_node_obj_names_embedded_slashes
+  call t_node_obj_fetch_existing
 
   call fruit_summary
   call fruit_finalize